+++ frontend/src/components/details/EmailContainer.tsx
